requirements.txt
# Core dependencies - Latest stable versions
tensorflow>=2.15.0
numpy>=1.24.0
Pillow>=10.0.0

# Development and testing dependencies
pytest>=7.4.0
pytest-cov>=4.1.0
black>=23.0.0
flake8>=6.0.0
mypy>=1.5.0

# Optional dependencies for enhanced functionality
matplotlib>=3.7.0
seaborn>=0.12.0
pandas>=2.0.0
scikit-learn>=1.3.0

# Data processing
opencv-python>=4.8.0
albumentations>=1.3.0

# Logging and monitoring
tensorboard>=2.15.0
wandb>=0.15.0

# Configuration management
pydantic>=2.0.0
python-dotenv>=1.0.0

# Additional utilities
tqdm>=4.65.0
requests>=2.31.0
